Move AfcError to error.rs
