+++ crates/aranya-client/src/afc.rs
@@ -45,123 +45,7 @@ use tokio::{
 };
 use tracing::{debug, error, instrument, warn};
 
-/// Possible errors that could happen when using Aranya Fast Channels.
-#[derive(Debug, thiserror::Error)]
-pub enum AfcError {
-    // Connection-related errors
-    /// Unable to bind a network addresss.
-    #[error("unable to bind address: {0}")]
-    Bind(io::Error),
-
-    /// DNS lookup failed.
-    #[error("DNS lookup failed: {0}")]
-    DnsLookup(io::Error),
-
-    /// Local address failure.
-    #[error("unable to get local address: {0}")]
-    RouterAddr(io::Error),
-
-    /// Unable to parse shm path.
-    #[error("unable to parse shared memory path: {0}")]
-    ShmPathParse(aranya_fast_channels::shm::InvalidPathError),
-
-    /// Unable to open the shm read state.
-    #[error("unable to open shared memory `ReadState`: {0}")]
-    ShmReadState(anyhow::Error),
-
-    /// Unable to accept a TCP stream.
-    #[error("unable to accept to TCP stream: {0}")]
-    StreamAccept(io::Error),
-
-    /// Unable to create a TCP stream.
-    #[error("unable to connect to TCP stream: {0}")]
-    StreamConnect(io::Error),
-
-    /// Unable to read from TCP stream.
-    #[error("unable to read from TCP stream: {0}")]
-    StreamRead(io::Error),
-
-    /// Unable to write to TCP stream.
-    #[error("unable to write to TCP stream: {0}")]
-    StreamWrite(io::Error),
-
-    /// Unable to shutdown TCP stream.
-    #[error("unable to shutdown TCP stream: {0}")]
-    StreamShutdown(io::Error),
-
-    /// Unable to get the remote peer's address.
-    #[error("unable to get remote peer's address: {0}")]
-    StreamPeerAddr(io::Error),
-
-    /// The stream was not found.
-    #[error("stream not found: {0}")]
-    StreamNotFound(SocketAddr),
-
-    // Protocol-related errors
-    /// Invalid AFC header.
-    #[error("invalid AFC header: {0}")]
-    InvalidHeader(#[from] aranya_fast_channels::HeaderError),
-
-    /// Invalid AFC magic.
-    #[error("invalid magic: {0}")]
-    InvalidMagic(u32),
-
-    /// Invalid AFC message.
-    #[error("invalid message: {0}")]
-    InvalidMsg(#[from] aranya_fast_channels::ParseError),
-
-    /// AFC message was replayed.
-    #[error("AFC message was replayed: {0}")]
-    MsgReplayed(String),
-
-    /// The message length prefix was larger than the maximum
-    /// allowed size.
-    #[error("message too large: {got} > {max}")]
-    MsgTooLarge { got: usize, max: usize },
-
-    /// Payload is too small to be ciphertext.
-    #[error("payload is too small to be ciphertext")]
-    PayloadTooSmall,
-
-    /// AFC message decryption failure.
-    #[error("decryption failure: {0}")]
-    Decryption(aranya_fast_channels::Error),
-
-    /// AFC message encryption failure.
-    #[error("encryption failure: {0}")]
-    Encryption(aranya_fast_channels::Error),
-
-    /// Serde serialization/deserialization error.
-    #[error("serialization/deserialization error: {0}")]
-    Serde(postcard::Error),
-
-    /// AFC version mismatch.
-    #[error("AFC version mismatch: got {actual:?}, expected {expected:?}")]
-    VersionMismatch { expected: Version, actual: Version },
-
-    // General errors
-    /// The channel was not found.
-    #[error("channel not found: {0}")]
-    ChannelNotFound(AfcId),
-
-    /// The 64-bit sequence number overflowed and the end of the channel was
-    /// reached. A new channel must be created.
-    ///
-    /// # Note
-    ///
-    /// This likely indicates that the peer manually set a very high sequence
-    /// number.
-    #[error("end of channel reached")]
-    EndOfChannel,
-
-    /// Some other error.
-    #[error("{0}")]
-    Other(#[from] anyhow::Error),
-
-    /// An internal bug was discovered.
-    #[error("internal bug: {0}")]
-    Bug(#[from] Bug),
-}
+use crate::error::AfcError;
 
 /// Data that can be polled from an Aranya Fast Channel.
 #[must_use]
